==> scripts/07_microbiome_msea.ipynb <==
# Microbe-Set Enrichment Analysis

Microbe-Set Enrichment Analysis[MSEA](https://www.nature.com/articles/s41598-020-78511-y#Sec8) is a host-microbe enrichment analysis package that checks for enriched microbes with known microbe-human gene interactions reported in literature. This needs a reference microbe set compendium (the one used here is provided by the authors) and a list of microbes of interest basically the one that comes up as significant in the differential abundance analysis.

# install msea package
#pip install msea

# Step 1: Read reference

Read in a GMT file with reference human-microbe.Here, the file provided by msea is taken into consideration which has a total of 1286 microbe-sets (gene: microbes) from published literature.

from pprint import pprint
import msea
from msea import SetLibrary
import pandas

# 0. load a reference microbe-set library from a GMT file
gmt_filepath = \
    'https://bitbucket.org/wangz10/msea/raw/aee6dd184e9bde152b4d7c2f3c7245efc1b80d23/msea/data/human_genes_associated_microbes/set_library.gmt'

# msea package provides a `read_gmt` function to parse a GMT file into a
# dictionary of sets
d_gmt = msea.read_gmt(gmt_filepath)
print('Number of microbe-sets:', len(d_gmt))
# Look at a couple of reference sets in the library
pprint(list(d_gmt.items())[:3])

# Step 2: msea.enrich
perform MSEA for a input microbe-set against the library of reference sets

#example
microbe_set_input = set(['Colwellia',
                         'Deinococcus',
                         'Idiomarina',
                         'Neisseria',
                         'Pseudidiomarina',
                         'Pseudoalteromonas'])
# this can be done using the `msea.enrich` function
msea_result = msea.enrich(microbe_set_input, d_gmt=d_gmt, universe=1000)
# check the top enriched reference microbe-sets
print(msea_result.head())

# Step 3: msea with adjustment
Perform MSEA with adjustment of expected ranks for reference sets sometimes certain reference microbe-sets in a library are more likely to be enriched by chance. We can adjust this by empirically estimating the null distributions of the ranks of the reference sets, then uses z-score to quantify if observed ranks are significantly different from the expected ranks.

# To do that, it is easier through the `SetLibrary` class.
set_library = SetLibrary.load(gmt_filepath)
# The `SetLibrary.get_empirical_ranks` method helps compute the expected
# ranks and store the means and stds of the ranks from the null
# distribution
set_library.get_empirical_ranks(n=20)
print(set_library.rank_means.shape, set_library.rank_stds.shape)
# After that, we can perform MSEA with this adjustment
msea_result_adj = set_library.enrich(microbe_set_input, adjust=True, universe=1000)
print(msea_result_adj.head())

#import math
math.log10(0.000030)*(-1.707623)

# Create msea_pml()

def msea_pml(microbe_set_input, gmt_filepath):
    set_library = SetLibrary.load(gmt_filepath)
    set_library.get_empirical_ranks(n=20)
    print(set_library.rank_means.shape, set_library.rank_stds.shape)
    msea_result_adj = set_library.enrich(microbe_set_input, adjust=True, universe=1000)
    print(msea_result_adj.head())
    return msea_result_adj

# Cancer.vs.Ctrl
## Up Microbes

microbe_set_input = set(['Fusobacterium',
                         'Dialister',
                         'Peptostreptococcus',
                         'Prevotella',
                         'Shewanella',
                         'Campylobacter',
                         'Veillonella',
                         'Neisseria',
                         'Porphyromonas',
                         'Solobacterium',
                         'Glaesserella',
                         'Haemophilus',
                         'Microlunatus',
                         'Bergeyella',
                         'Lachnoanaerobaculum',
                         'Bibersteinia',
                         'Aggregatibacter',
                         'Obesumbacterium',
                         'Lancefieldella',
                         'Centipeda',
                         'Snodgrassella',
                         'Riemerella',
                         'Schaalia',
                         'Plesiomonas',
                         'Gemella',    
                         'Simonsiella',
                         'Leptotrichia'])
can_ctrl_up = msea_pml(microbe_set_input=microbe_set_input, gmt_filepath=gmt_filepath)
can_ctrl_up.to_csv('/restricted/projectnb/montilab-p/projects/oralcancer/pml/results/mmkhan/msea_cancer_ctrl_up.csv')

## Dn Microbes

microbe_set_input = set(['Gardnerella',
                         'Limnohabitans',
                         'Kocuria',
                         'Ralstonia',
                         'Thalassospira',
                         'Agrobacterium', 
                         'Streptomyces',
                         'Delftia', 
                         'Bradyrhizobium',
                         'Streptomyces',
                         'Malassezia',
                         'Pseudoglutamicibacter',
                         'Rhodoluna',
                         'Tepidimonas'
                        ])
                            
can_ctrl_dn = msea_pml(microbe_set_input=microbe_set_input, gmt_filepath=gmt_filepath)
can_ctrl_dn.to_csv('/restricted/projectnb/montilab-p/projects/oralcancer/pml/results/mmkhan/msea_cancer_ctrl_dn.csv')

# PML.vs.Ctrl
## Up Microbes
Combining both HkNR+Dys microbes

microbe_set_input = set(['Shewanella',
                         'Aggregatibacter',
                         'Haemophilus',
                         'Mannheimia', 
                         'Veillonella',
                         'Dialister',
                         'Fusobacterium',
                         'Neisseria',
                         'Riemerella',
                         'Leptotrichia',
                         'Campylobacter', 
                         'Peptostreptococcus',
                         'Porphyromonas',
                         'Streptococcus',
                         'Plesiomonas',
                         'Bibersteinia', 
                         'Prevotella',
                         'Gemella',
                         'Microlunatus',
                         'Novosphingopyxis',
                         'Marivita',
                         'Oceanivirga',
                         'Simonsiella',
                         'Bergeyella',
                         'Snodgrassella',
                         'Solobacterium',
                         'Lachnoanaerobaculum',
                         'Mannheimia',
                         'Schaalia',
                         'Canicola' 
                         ])
pml_ctrl_up = msea_pml(microbe_set_input=microbe_set_input, gmt_filepath=gmt_filepath)
pml_ctrl_up.to_csv('/restricted/projectnb/montilab-p/projects/oralcancer/pml/results/mmkhan/msea_pml_ctrl_up.csv')

## Dn Microbes

# not found: Curvibacter;Centipeda;Lancefieldella;Catonella;Rhodoluna;Lacrimispora;Cronobacter
microbe_set_input = set(['Agrobacterium',
                         'Selenomonas',
                         'Rhodoluna'
                        ])
pml_ctrl_dn = msea_pml(microbe_set_input=microbe_set_input, gmt_filepath=gmt_filepath)
pml_ctrl_dn.to_csv('/restricted/projectnb/montilab-p/projects/oralcancer/pml/results/mmkhan/msea_pml_ctrl_dn.csv')
